Pretend to only support version 0.4 of the bootloader.
Recent versions of the FOTA android app require an init packet to be
present. This requirement seems to depend on the version advertised by
the bootloader.
With this simple change, we allow users to update their devices without
having to bundle an init packet every time, which would be quite
annoying when prototyping.